roles y pooliticas de acceso

+++ app/Http/Controllers/Admin/RolesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Http\Requests\SaveRolesRequest;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
@@ -16,6 +17,7 @@ class RolesController extends Controller
      */
     public function index()
     {
+        $this->authorize('view', new Role);
         return view('admin.roles.index', [
             'roles' => Role::all()
         ]);
@@ -28,8 +30,9 @@ class RolesController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', $role = new Role);
         return view('admin.roles.create',[
-                'role' => new Role,
+                'role' => $role,
                 'permissions' => Permission::pluck('name', 'id'),
             ]);
     }
@@ -40,29 +43,14 @@ class RolesController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(SaveRolesRequest $request)
     {
-        $data = $request->validate([
-            'name' => 'required|unique:roles',
-            'guard_name' => 'required'
-        ]);
-
-        $role = Role::create($data);
+        $this->authorize('create', new Role);
+        $role = Role::create($request->validated());
 
         $role->givePermissionTo($request->permissions);
 
         return redirect()->route('admin.roles.index')->withFlash('El rol fue creado correctramente');
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
     }
 
     /**
@@ -73,6 +61,7 @@ class RolesController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('delete', $role);
         return view('admin.roles.edit', [
             'role' => $role,
             'permissions' => Permission::pluck('name', 'id'),
@@ -86,13 +75,10 @@ class RolesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Role $role)
+    public function update(SaveRolesRequest $request, Role $role)
     {
-        $data = $request->validate([
-            'display_name' => 'required',
-        ]);
-
-        $role->update($data);
+        $this->authorize('update', $role);
+        $role->update($request->validated());
 
         $role->permissions()->detach();
 
@@ -108,8 +94,11 @@ class RolesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Role $role)
     {
-        //
+
+        $this->authorize('delete', $role);
+        $role->delete();
+        return redirect()->route('admin.roles.index')->withFlash('El role fue eliminado');
     }
 }
